Create SingleBook.js
